+++ src/project_sniffer/reading/__init__.py
@@ -1,4 +1,5 @@
 from project_sniffer.reading.file_reader import (
+    DEFAULT_MAX_SOURCE_BYTES,
     read_manifest_files,
     read_scanned_file,
 )
@@ -9,6 +10,7 @@ from project_sniffer.reading.models import (
 
 
 __all__ = [
+    "DEFAULT_MAX_SOURCE_BYTES",
     "FileReadResult",
     "FileReadStatus",
     "read_manifest_files",